avoid to read a configuration ack after start acquisition
After the acquisition was started, using the CS command, the device leaves the
configuration mode. Therefor no further configuration ack will be send.
My guess why this was working most of the time before is that there might have been
enough ack's in the buffer due to the carriage return's send in setConfigurationMode